performance-timeline/test_performance_timeline_functions.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title></title>
<link rel="author" title="W3C" href="http://www.w3.org/" />
<link rel="help" href="http://www.w3.org/TR/performance-timeline/"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="stylesheet" href="/resources/testharness.css" />
<script>
test(function() {
assert_equals(typeof window.performance.getEntries, "function");
}, "Verify getEntries exists and is a function");
test(function() {
assert_equals(typeof window.performance.getEntriesByName, "function");
}, "Verify getEntriesByName exists and is a function");
test(function() {
assert_equals(typeof window.performance.getEntriesByType, "function");
}, "Verify getEntriesByType exists and is a function");


</script>
</head>
<body>
<h1>Description</h1>
<p>This test verifies the existence of getEntries, getEntriesByType and getEntriesByName functions that are defined in the Performance Timeline specification.</p>

<div id="log"></div>

</body>
</html>
